Update config/application.config.php
Prevent adding a module if it's already in the list.
Because on some systems it's not possible to define environment variables, you should be able to add the module to the normal module list (defined at the top of this file). Problem occures when you work with different systems for dev/test/prod etc. and want to share the same code base.