Evergreen ILS holdings improvement
The Evergreen driver was passing in the record ID but returning the copy
of the same ID, meaning that not only were the wrong copies being
returned for Evergreen holdings, but only one copy maximum was being
returned.

In addition, rather than interpolating strings into SQL for a possible
SQL injection attack, use parameter markers for the prepared statements
and bind our variables to those.

Signed-off-by: Dan Scott <dan@coffeecode.net>