Symphony: check both unseen, total renewals left
Add a check for total number of renewals remaining when checking whether an
item can be renewed.

The total number of renewals remaining (the values renewalsRemaining and
renewalsRemainingUnlimited) govern, even if unseen renewals remain, or are
unlmited. That is, Symphony treats unseen renewals as limiting the conditions
in which a renewal can take place, not a separate type of renewal that can be
performed without the item in hand.

Refer to SirsiDynix KB #131608 for additional details.