Allow setting ZF2 baseUrl via environment variable
ZF2 seems to be unable to detect the base url using a proper virtual host
configuration for dynamically created views via the admin interface. Thus we
need to set the baseUrl using an environment variable in the virtual host
configuration.

This may be due to a bug in ZF2 or simply because the framework is heavily
built around the idea, that public resides in the DocumentRoot for the virtual
host. Nevertheless, this idea allows us to skirt around the fact that ZF2 fails
at detecting the path correctly.

Also, I've added PHPStorm project file directory to .gitignore