Symphony: bill creation date never returned
The bill creation date is returned in the top level FeeInfo object returned
by Web Services, unlike the Checkout and Due Dates, which are returned
in the FeeItem child object, which described the item (book, etc.)
on which the fee was incurred (if appropriate). The code assumed
that all dates would be in the FeeItem object.

Resolves VUFIND-1115.