Fix next/prev scrolling for author facet searches
When running a keyword search, then choosing an author hyperlink from
the
results to run an author facet search, the next/previous links in the
individual
results would be based on the initial keyword search, not the facet
search.

This was for two reasons:
- the ResultScroller was not active in the AuthorController, which
handles
author facet searches
- the ResultScroller retrieves the current search from the session to
create
the scrolling links, and author facet searches were configured to not
be saved in the session

Correct these two issues in AuthorController, but only if
next_prev_navigation
is enabled. Because AuthorController also handles SolrAuthorFacet
searches
(searches that produce a list of authors, not a list of titles by a
specific author)
and those results cannot be iterated through by a ResultScroller, we
also have
to make an additional fix in the ResultScroller plugin to handle the
case
where SolrAuthorFacet result is passed to create the ResultScroller, and
prevent it from being instantiated.