[deb] Support for php7 and mariadb
As discussed on https://sourceforge.net/p/vufind/mailman/message/35002108/

This is the main change i would like to make in time to 3.0 release. The packaging process needs other changes that i'm planning to discuss and make (if you're willing) but they're not so much important as this one. This will make vufind compatible with php7 and mariadb de facto, because as long as vufind itself already support them its deb package don't.

I would like to point out that this changes won't break the previous behaviour, it will just introduce new alternative dependencies based on virtual packages that can either provide php5 or php7, i did not test them yet, i'm planning to virtualize a debian sid to do so but it's very unlikelly that something is wrong and if it is, it would just break the php7 use case.

How are you guys on the release schedule? I'm planning to make the test at about this friday, but i understand that i may be pushing you guys to close to the deadline and if that's the case i can hurry up and test it tomorrow or wednesday.