Set pickupLocation=homebranch(es) when neither defaultPickupLocation nor pickupLocation[] is set in KohaILSDI.ini
When defaultPickupLocation AND pickupLocationp[] is empty in KohaILSDI.ini and holds are enabled, the user would get an error.
In that case, as a last resort, it is relatively safe to use the homebranch location as a pickup location.

When in item level holds: get the one location the selected item is available at.
When in title level holds: get the unique locations the selected title is available at.