Allow sending records by sms to be disabled in config
Currently, sending record results by sms ("Text this" feature) is
always shown in the record page, even though it may not be applicable
to some installations.

Although still left enabled by default, with this patch, the admin
has the option to suppress the sms link in the record page, if needed.

The [Mail] section in main config file has been chosen to hold the new
parameter because sending sms messages uses the underline mailer
infrastructure.