Rename $currentDir to $baseDir and update docs.
This change is introduced as part of adding platform (Heroku) support. Heroku uses a directory in /tmp to build the project and then moves it to /app on completion. Simultaneously, install.php uses a variable ($baseDir) in a way that does not distinguish between "current directory of install.php" and "VuFind base directory," because up until this point they were absolutely the same. In a platform Heroku, though, it now means that install.php will specify a /tmp directory in any generated config files while in fact needing to specify /app.

This commit renames the $baseDir variable within the install.php script to $currentDir. That way, each use of that variable can offer a clear decision point as to whether the current directory is the truly intended use or if there should be some other value. This is more of a quality of life change for a developer working within install.php when platforms are supported, especially as platform support expands outside just the Apache config.

The obvious hacky "fix" is to just override $baseDir, but that increases the complexity of what that variable means in a procedural script and makes understanding what's going on harder. I think it's more clear to the developer if I make the changes above and turn it into a decision between $currentDir and $someOtherDir.