Require --non-interactive and suppress directions with --platform.
We require --non-interactive with --platform because 1) automation is a required paradigm for platform support and 2) some capabilities aren't yet truly supported with --platform, so it's safer to force the use of defaults and minimize unexpected outcomes. Additionally, we don't want to display unhelpful or incorrect instructions when --platform is specified, so we suppress them.