Mimic vendor lib structure in solr container.
The biblio core config expects `solr/vendor/contrib` to exist two
levels above it. The official solr container does not have that
structure (it's missing `vendor`). So we create that directory
and symlink the contrib folder on container init to mimic the
expected structure. Changing the biblio core config instead might
result in confusion/disruption of non-Dockerized development, so I
did this instead.