Conditionally run composer install on container launch.
In a development setup, we mount our local VuFind workspace in
place of the files baked into the Docker image, because we want our
local changes to affect the running container. Unfortunately, this
also means the vendor folder in the local workspace overrides that
which is baked into the Docker image. If there's anything lingering
in there, it can upset VuFind. So what we need to do is run
composer install from within the container to cause the vendor
folder in our local workspace to contain the right set of
dependencies. The first time the container runs, this composer
install step may take a while. Future runs should be quick, since
nothing will need to be done. We should be able to improve this
performance by directing phing to not install solr when docker is
involved.