Don't lose followup URL due to mistyped credentials (#1269)
If a user accidentally mistypes their login credentials, they are not returned to the page they were on after successfully logging in via ILS Authentication. There already exists a similar fix for when a non-logged-in user goes to MyResearch/Home (which initiates the login challenge page). This bug fix adds one for MyResearch/UserLogin, too, which is linked to in themes/bootstrap3/templates/header.phtml

    Note: Links to MyResearch/UserLogin can have parameters, e.g., UserLogin?method=ILS,
    which is why I perform a substring test here:

        if (0 === strpos($refererNorm, $mulNorm)) {

     instead of testing for equality:

        if ($refererNorm === $mulNorm) {