KohaILSDI Fix to return all recieved serial issues
Fix in KohaILSDI getPurchaseHistory function in order to get as
a result an array of ALL recieved issues of a serial.
(This behavior better matches the function documentation and is
consisntent with getPurchaseHistory results from other ILS drivers)
In case one wants to limit the number of displayed issues, this
should be ideally handled in the theme template.