Alma: Fix date and status display for the holds list. (#1501)
As far as I can see there's no way to determine if a request is in transit. I'm open to suggestions if e.g. @betullam knows more.

I suspect the same changes should be done to the other two methods, but I don't currently have a way of testing it.