Make form sender fields overridable.
This makes it possible to e.g. define the field positions on the form and add additional fields like a phone number in the correct position. These and other reserved fields are now documented in FeedbackForms.yaml.

The email field now accepts the maximum length that an email address can be.