Fix: Construction of Query String for EDS Autocomplete Requests
EDS autocomplete fails within the old code as the json that was enclosed in the filters parameters wasn't fully URLEncoded.

This changes create an array of parameters and then uses a built in PHP function to correctly build the query string. Testing with this amended function worked correctly locally and for a German VuFind User