Feature: get correct dueDate for Folio
Folio requires a third call to obtain the dueDate. The /circulation/loans call will potentially return an array of loans, the the currently active loan is the one that does not have a returnDate.

Due to the third API call a Folio.ini configurable option has been introduced to limit the number of "dueDates" to fetch (limit to 5 by default, to avoid too long wait times for books with many checked out items.

Additionally Folio.ini has configurable values as to whether to show dueDates at all and if a time should be shown. There is significant room for improvement here.