[VUFIND-1597] tweak selection of format for generating OpenURLs
This patch adjusts the algorithm for selecting which OpenURL
format to use for a given record.

It addresses cases where the record format ends up as Serial
due to vagaries of the cataloging of the original MARC record
but that should nonetheless have any available ISSNs included
in their OpenURL. It also attempts to avoid sending both ISBN
and ISSN for titles that are part of monographic series.

Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>