Add support for optional "required by" date for holds. (#2740)
"requiredByDateOptional" can be used when the user needs to be able to choose a last effective date for a hold, but is not required to do so. Additionally "defaultRequiredDate" can now be set to no value or empty string for no default. Support for unspecified "required by" date is included for Demo, KohaRest and SierraRest drivers.