Fix visual view link and jump to record handling.
If user jumps to visual view from e.g. the second page, we need to ensure the view url doesn't have a page number. This avoids drilling into visual results from displaying e.g. results 21 to 2 of 2.
Also adds a check that a record was returned before trying to jump to it.