Fix login followup URL for off-site auth
When using a link to an off-site auth system (single-sign-on)
in the login lightbox, the followup URL was not being preserved.
Lightbox handled reloading the page for direct logins, but that
doesn't help if the user leaves the site to authenticate.

This patch makes it so the login followup URL is always set,
but then clears that URL on login if the user is inside
the lightbox.