Display custom hidden filters as a temporary search tab. (#2908)
Previously the default tab was displayed as active when current hidden filters did not match any of the tabs, which was quite misleading and confusing. This change makes it possible to display a temporary custom tab in this case, or optionally not display any tab as active by emptying the translation.