Add support for Primo REST API. (#3255)
Deprecates the legacy brief search API support and adds support for the current REST API for Primo search. Also includes a fix for exclude filters clashing with normal filters for the same field. Processing of the format field is moved to record driver to avoid duplication (and doing it in the right place).