Fix LocalFile.php to properly handle %size%
When using LocalFile in the [Content] area in config.ini, you have the option to use %size% token to distinguish directories based on different sizes (small/medium/large).

Up till now, the %size% token was not handled and translated to one of the 3 allowed values, resulting in a filename that contained the %size% string in it, thus not pointing to the proper directory in the filesystem.

This fix remedies that (rare) issue.