Fix first/last navigation. (#3404)
Separates record page first/last navigation setting from backend support for first/last navigation by improved naming of the variable and method. Also adds flag about support for first/last navigation in results to the Options class so that it works properly with JS results and doesn't need to be set in a backend-specific template.