Defer login token update to end of request and tweaks emails.
It was possible for the processing of the request to fail after the login token was updated, so the new token was never communicated to user's browser. This change defers the token update until other processing is complete. It also softens the email messages a bit since there's always a potential for technical issues instead of an attack, and a detected issue means the login was not successful.
Also the suggestion to change password has been removed since it does nothing if cookies get stolen, and is not a useful suggestion with email authentication.
A configuration option to disable emails has been added so that they can be easily disabled if it seems that they cause too much trouble.