Extend itemLimit config to KohaRest driver and ini file
The itemLimit config is available with the Alma driver and it would be
nice if the KohaRest driver could also use it. Especially, given the
Koha-plugin-rest-di plugin will accept a 'limit' param.

To give more configurability we have enhanced the itemLimit config so
that libraries can define different numbers of items to be displayed for
different Koha biblio-level item types - which are defined in the 942$c
subfield in Koha.

Sponsored-by: Auckland University of Technology, New Zealand