VUFIND-1710 Add configurable separators for topics in alphabrowse
Currently, subjects are shown differently between the record view in
VuFind and the topic browse.  In the record view you see delimiters
like this:

    Morris, William, > 1834-1896 > Criticism and interpretation.

But in the topic browse, the fields are just joined with spaces:

    Morris, William, 1834-1896 Criticism and interpretation.

This commit:

  * Adds a UTF-8 separator to delimit terms within topics in the Solr
    index

  * Adjusts the topic_browse definition to use a new TopicNormalizer,
    which strips out these new delimiters when building the sort key.

  * Changes the frontend of the AlphaBrowse to insert a
    user-configurable delimiter, shown when browsing topics.

Note: since Normalizer classes currently live in the
vufind-browse-handler, there will be a corresponding commit there to
add TopicNormalizer.