Fix CurrencyFormatter and remove invalid tests.
We used to rely on setlocale returning the previously set locale value, but this is not the case with PHP 8.4 (and isn't documented either). As far as I can see there's no way for the test to set the default locale anymore, so the tests relying on it don't work properly anymore with PHP 8.4.