[release-10.1] Fix form validation (#4227)
The old validation mechanism was not compatible with Bootstrap 5 and didn't work properly with Bootstrap 3 either. The new uses the JS validation API. It is theme-independent and less intrusive with the caveat that the validation error messages are displayed in a browser-specific style.

(cherry picked from commit 9f0985eec977696c81f1d020dcdd56a33932c7a6)