introducing hiddenRTACId
hiddenId is used for functions beyond availability checking.

EDS > result-list.phtml introduces a new value "hiddenRTACId". For EDS this will always be present and either have the value of the appropriate id or "no_rtac"

if hiddenRTACId is present, it will replace the hiddenIdEl for availability checks.

For any id equalling "no_rtac" and early exit in the function avoid superfluous availability checks.

this approach caused an inconvenience in "lists" a list-entry.phtml was added to RecordDriver/EDS/ to also introduce the hiddenRTACId