removing "getFromatList"
The format is shown below the publication type icon to the left of the record. While this is different to the rest of VuFind this change has no functional significance to the focus of the PR and I am thus removing it, as I had previously with the misunderstood getLabelList function call.

Thank you for pointing out this