Remove umask from cache settings (#4906)
The umask is a global setting which does not only affect cache related files
and directories, but any other files and directories, too.

Setting umask outside of VuFind or fixing directory permissions after the
directories were created are alternatives which are easy to implement for
installations which want a special umask.

Signed-off-by: Stefan Weil <sw@weilnetz.de>